tools/requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
googleapis-common-protos==1.60.0
grpcio==1.56.2
idna==3.4
iniconfig==2.0.0
#packaging==23.1
pluggy==1.2.0
protobuf==4.24.3
#pycparser==2.21
PyJWT==2.8.0
#pytest==7.4.0
#python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
urllib3==2.0.4
yandexcloud==0.248.0
